% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianNetworks-package.R
\docType{package}
\name{BayesianNetworks-package}
\alias{BayesianNetworks}
\alias{BayesianNetworks-package}
\title{BayesianNetworks: Bayesian Modelling of Bipartite Networks}
\description{
Bayesian modelling of bipartite network structure, following the approach of Young et al. \doi{10.1038/s41467-021-24149-x}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Pakillo/BayesianNetworks}
  \item Report bugs at \url{https://github.com/Pakillo/BayesianNetworks/issues}
}

}
\author{
\strong{Maintainer}: Francisco Rodriguez-Sanchez \email{f.rodriguez.sanc@gmail.com} (\href{https://orcid.org/0000-0002-7981-1599}{ORCID})

}
\keyword{internal}
