% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianWebs-package.R
\docType{package}
\name{BayesianWebs-package}
\alias{BayesianWebs}
\alias{BayesianWebs-package}
\title{BayesianWebs: Bayesian Modelling of Bipartite Networks}
\description{
Bayesian modelling of bipartite network structure, following the approach of Young et al. \doi{10.1038/s41467-021-24149-x}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pakillo.github.io/BayesianWebs/}
  \item Report bugs at \url{https://github.com/Pakillo/BayesianWebs/issues}
}

}
\author{
\strong{Maintainer}: Francisco Rodriguez-Sanchez \email{f.rodriguez.sanc@gmail.com} (\href{https://orcid.org/0000-0002-7981-1599}{ORCID})

Other contributors:
\itemize{
  \item William K. Petry \email{wpetry@ncsu.edu} (\href{https://orcid.org/0000-0002-5230-5987}{ORCID}) [contributor]
}

}
\keyword{internal}
