% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_model.R
\name{check_model}
\alias{check_model}
\title{Check fitted model}
\usage{
check_model(fit = NULL)
}
\arguments{
\item{fit}{A fitted model, as obtained from \code{fit_model()}.}
}
\value{
Model checks on console and graphical window.
}
\description{
Check fitted model
}
\examples{
data(web)
dt <- prepare_data(mat = web, sampl.eff = rep(20, nrow(web)))
fit <- fit_model(dt, refresh = 0)
check_model(fit)
}
