% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Fit model}
\usage{
fit_model(data = NULL, ...)
}
\arguments{
\item{data}{A named list containing the required data, as obtained from
\code{prepare_data()}.}

\item{...}{Further arguments for \code{cmdstanr::sample()}, like \code{iter_warmup},
\code{iter_sampling}, or \code{thin}, among others.}
}
\value{
A fitted model (\code{cmdstanr::CmdStanMCMC()} object).
}
\description{
Fit model
}
\examples{
data(web)
dt <- prepare_data(mat = web, sampl.eff = rep(20, nrow(web)))
fit <- fit_model(dt)
}
