% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_posterior.R
\name{get_posterior}
\alias{get_posterior}
\title{Get posterior values}
\usage{
get_posterior(
  fit = NULL,
  data = NULL,
  param = c("all", "connectance", "preference", "plant.abund", "animal.abund",
    "int.prob", "link")
)
}
\arguments{
\item{fit}{Fitted model (from \code{fit_model()})}

\item{data}{Data list (from \code{prepare_data()})}

\item{param}{character. Name of the parameter to retrieve the posterior samples.}
}
\value{
A data frame
}
\description{
Get posterior values
}
\examples{
data(web)
dt <- prepare_data(mat = web, sampl.eff = rep(20, nrow(web)))
fit <- fit_model(dt, refresh = 0)
connectance <- get_posterior(fit, dt, param = "connectance")
connectance

int.prob <- get_posterior(fit, dt, param = "int.prob")
int.prob
int.prob |> tidybayes::mean_qi()  # mean edge probability

# plot
int.prob |>
  tidybayes::mean_qi() |>
  dplyr::select(Plant, Animal, int.prob) |>
  network.tools::plot_web_heatmap(int.var = "int.prob", sort = FALSE)

 # compare with Visits
 web |>
   network.tools::wide2long() |>
   network.tools::plot_web_heatmap(zero.na = FALSE, sort = FALSE)

# all posteriors
get_posterior(fit, dt, param = "all")
}
