% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_seed.R
\name{get_seed}
\alias{get_seed}
\title{Get seed used to fit a model}
\usage{
get_seed(fit = NULL)
}
\arguments{
\item{fit}{A fitted model}
}
\value{
A number
}
\description{
Get seed used to fit a model
}
\examples{
data(web)
dt <- prepare_data(mat = web, sampl.eff = rep(20, nrow(web)))
fit <- fit_model(dt)
get_seed(fit)
}
