% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_there_link.R
\name{is_there_link}
\alias{is_there_link}
\title{Is there a link?}
\usage{
is_there_link(df = NULL)
}
\arguments{
\item{df}{A dataframe containing posterior interaction probabilities, as
generated from \code{get_posterior()}}
}
\value{
A data frame with a new column \code{link} taking values of 0 (no link)
and 1 (link).
}
\description{
Is there a link?
}
