% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{optimal_bayes}
\alias{optimal_bayes}
\title{Fits a very simple "Optimal Bayes" classifier.}
\usage{
optimal_bayes(formula, data)
}
\arguments{
\item{formula}{input formula with the hypothesis variable on the left-hand
side and the evidence variables on the right-hand side.}

\item{data}{matrix or dataframe with categorical (character/factor/logical) predictors.}
}
\value{
returns an object containing estimated probabilities for each possible combination
 of the values seen in the input data.
}
\description{
Fits a very simple "Optimal Bayes" classifier.
}
