% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare the data for modelling}
\usage{
prepare_data(mat, plant_effort = NULL, plant_abun = NULL, animal_abun = NULL)
}
\arguments{
\item{mat}{A matrix containing quantitative counts (not qualitative or binary)
of observed interactions (e.g. visits to flowers, number of fruits
consumed per plant or species). Plants must be in rows, Animals must be in columns.}

\item{plant_effort}{A numeric vector with the sampling effort (e.g. observation hours)
spent on each plant.}

\item{plant_abun}{(optional) A numeric vector with the relative or absolute abundances of each plant.}

\item{animal_abun}{(optional) A numeric vector with the relative or absolute abundances of each animal.}
}
\value{
A named list containing data to be passed to the bipartite network model.
}
\description{
Prepare the data for modelling
}
\examples{
data(web)
prepare_data(web, plant_effort = rep(20, nrow(web)))
}
