% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{simulate.sprinkler}
\alias{simulate.sprinkler}
\title{Simulates data for the "Sprinkler" example.}
\usage{
\method{simulate}{sprinkler}(N = 1000)
}
\arguments{
\item{N}{number of data points to generate.}
}
\value{
Returns a data frame containing the simulated data.
}
\description{
Simulates data for the "Sprinkler" example.
}
