% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{simulate_sprinkler}
\alias{simulate_sprinkler}
\title{Simulates data for the "Sprinkler" example.}
\usage{
simulate_sprinkler(N = 1000)
}
\arguments{
\item{N}{number of data points to generate.}
}
\value{
returns a data frame containing the simulated data.
}
\description{
Simulates data for the "Sprinkler" example.
}
