% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stancode.R
\name{stancode}
\alias{stancode}
\title{Make Stan code}
\usage{
stancode(
  plant_effort = c("param", "data"),
  animal_pref = c("fixed", "varying"),
  plant_abun = c("estimate", "relative", "absolute"),
  animal_abun = c("estimate", "relative", "absolute")
)
}
\value{
A string containing a Stan model.
}
\description{
Make Stan code
}
\details{
This function is designed to be used internally. See \code{fit_model} for documentation
of arguments.
}
\keyword{internal}
