% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{web}
\alias{web}
\title{Plant-pollinator network}
\format{
\subsection{\code{web}}{

A numeric (integer) matrix with 8 rows (representing plants) and 21 columns
(representing animals)
}
}
\source{
Kaiser-Bunbury, C., Mougal, J., Whittington, A. et al.
Ecosystem restoration strengthens pollination network resilience and function.
Nature 542, 223–227 (2017). https://doi.org/10.1038/nature21071
}
\usage{
web
}
\description{
An example bipartite network of 8 plant species and 21 pollinators,
from \href{https://doi.org/10.1038/nature21071}{Kaiser-Bunbury et al. 2017}.
}
\keyword{datasets}
