% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Fit bipartite network model}
\usage{
fit_model(
  data = NULL,
  plant_effort = c("param", "data"),
  animal_pref = c("fixed", "varying"),
  plant_abun = c("param", "data"),
  animal_abun = c("param", "data"),
  beta = 0.01,
  alpha_sigma = 1,
  alpha_tau = 1,
  ...
)
}
\arguments{
\item{data}{A named list containing the required data, as obtained from
\code{\link[=prepare_data]{prepare_data()}}.}

\item{plant_effort}{How to model the sampling effort allocated to each plant. If unknown, \code{param} (default) estimates a single parameter during the modelling step that applies to all plants, \eqn{C}. Alternatively, "data" allows measured values of sampling effort to be supplied in \code{data}, \eqn{C_i}.}

\item{animal_pref}{How to model the increase in visits of animals to plants with which they share a link relative to unlinked plants. \code{fixed} (default), fits a single preference parameter for all animals during the modelling step, \eqn{r}. \code{varying} fits a separate preference parameter for each animal type, \eqn{r_i}.}

\item{plant_abun}{How the model treats the abundances of interacting plants. \code{param} (default) treats these abundances as a parameter that the model estimates, \eqn{sigma_i}. \code{data} treats these abundances as measured and requires the user to supply the absolute abundances of each plant in \code{data}.}

\item{animal_abun}{How the model treats the abundances of interacting animals. \code{param} (default) treats these abundances as a parameter that the model estimates, \eqn{sigma_i}. \code{data} treats these abundances as measured and requires the user to supply the absolute abundances of each animal in \code{data}.}

\item{beta}{Rate parameter of exponential prior on the preference parameter or parameters \eqn{r} or \eqn{r_i}. The default, 0.01, is suitable for small counts and should be increased for larger counts. Use \code{plot_prior} to visualize.}

\item{alpha_sigma}{Concentration parameters of Dirichlet prior on the abundances of interacting plants, \eqn{sigma_i}, when \code{plant_abun = "param"}. Ignored otherwise. Defaults to 1.}

\item{alpha_tau}{Concentration parameters of Dirichlet prior on the abundances of interacting plants, \eqn{tau_i}, when \code{animal_abun = "param"}. Ignored otherwise. Defaults to 1.}

\item{...}{Further arguments passed to \code{\link[cmdstanr:model-method-sample]{cmdstanr::sample()}}, like \code{iter_warmup},
\code{iter_sampling}, or \code{thin}, among others.}
}
\value{
A fitted model (\code{\link[cmdstanr:CmdStanMCMC]{cmdstanr::CmdStanMCMC()}} object).
}
\description{
Fit bipartite network model
}
\details{
This function allows fits generalized variations of the bipartite network structure model of Young et al. 2021 \href{https://doi.org/10.1038\%2Fs41467-021-24149-x}{\url{doi:10.1038/s41467-021-24149-x}}.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
data(web)
dt <- prepare_data(mat = web, plant_effort = rep(20, nrow(web)))
# fit the model used by Young et al. 2021
fit <- fit_model(dt, plant_effort = "param", animal_pref = "fixed",
                 plant_abun = "param", animal_abun = "param")
\dontshow{\}) # examplesIf}
}
