% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Fit model}
\usage{
fit_model(data = NULL, model = "default", beta = 0.01, ...)
}
\arguments{
\item{data}{A named list containing the required data, as obtained from
\code{prepare_data()}.}

\item{model}{character. Path to file describing the Stan model. Use in case
you want to use a modified Stan model. Otherwise using "default" model provided
with the package.}

\item{beta}{Rate of exponential prior on \code{r} (preference) parameter.
Default beta is 0.01. Increase it if you have large count numbers.
(Could use \code{plot_prior()} to examine the resultant prior.}

\item{...}{Further arguments for \code{cmdstanr::sample()}, like \code{iter_warmup},
\code{iter_sampling}, or \code{thin}, among others. It is recommended to increase the
number of iterations (e.g. iter_sampling = 10000).}
}
\value{
A fitted model (\code{cmdstanr::CmdStanMCMC()} object).
}
\description{
Fit model
}
\examples{
data(web)
dt <- prepare_data(mat = web, sampl.eff = rep(20, nrow(web)))
fit <- fit_model(dt)
}
