% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_counts_pred.R
\name{plot_counts_pred}
\alias{plot_counts_pred}
\title{Plot heatmap of predicted counts}
\usage{
plot_counts_pred(pred_df = NULL, .width = 0.95, .point = mean, ...)
}
\arguments{
\item{pred_df}{A data frame containing the predicted counts, as generated by
\code{\link[=predict_counts]{predict_counts()}}}

\item{.width}{probability cutoff to determine the width of the resulting interval (default 0.95)}

\item{.point}{point summary function, which takes a vector and returns a single value, e.g. \code{\link[=mean]{mean()}}, \code{\link[=median]{median()}}, or \code{\link[=Mode]{Mode()}}}

\item{...}{Further arguments for \code{\link[network.tools:plot_web_heatmap]{network.tools::plot_web_heatmap()}}.}
}
\value{
A ggplot object
}
\description{
Plot heatmap of predicted counts
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
data(web)
dt <- prepare_data(mat = web, plant_effort = rep(20, nrow(web)))
fit <- fit_model(dt, refresh = 0)
pred_df <- predict_counts(fit, dt)
plot_counts_pred(pred_df)
plot_counts_pred(pred_df, sort = FALSE)
\dontshow{\}) # examplesIf}
}
