% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prior.R
\name{plot_prior}
\alias{plot_prior}
\title{Plot prior distribution for r (preference) parameter}
\usage{
plot_prior(beta = 0.01)
}
\arguments{
\item{beta}{A number > 0. Rate of the exponential distribution.}
}
\value{
A plot
}
\description{
The r (preference) parameter in Young et al. model takes a prior
exponential distribution with rate = beta. Use this function to
visualise the prior distribution of \code{r} given the chosen beta.
}
\examples{
plot_prior(beta = 0.01)
plot_prior(beta = 0.001)
}
